// src/app/layout/models/notifications.models.ts
export class Notifications{
    owner?: string;
    type?: string;
    content?: string;
    campaign_name?: string;
    campaign_id?: number;
    createdAt?: any;
    isArchived?: boolean;
}

export class Notifications_Accounts_Link{
    owner?: string;
    type?: string;
    content?: string;
    campaign_name?: string;
    campaign_id?: number;
    createdAt?: any;
    isArchived?: boolean;
}